update reservation, extra payment and transaction page design
