+++ react/src/dashboard/components/pages/reservations-view/elements/panes/ReservationExtraPaymentDetailsPane.tsx
@@ -1,19 +1,20 @@
 import React, { useCallback, useMemo } from 'react';
-import Organization from '../../../../props/models/Organization';
-import Reservation from '../../../../props/models/Reservation';
-import ExtraPayment from '../../../../props/models/ExtraPayment';
-import KeyValueItem from '../../../elements/key-value/KeyValueItem';
-import useSetProgress from '../../../../hooks/useSetProgress';
-import usePushSuccess from '../../../../hooks/usePushSuccess';
-import useProductReservationService from '../../../../services/ProductReservationService';
-import useOpenModal from '../../../../hooks/useOpenModal';
-import useEnvData from '../../../../hooks/useEnvData';
-import useTranslate from '../../../../hooks/useTranslate';
-import usePushApiError from '../../../../hooks/usePushApiError';
-import Label from '../../../elements/image_cropper/Label';
-import ModalReservationExtraPaymentRefund from '../../../modals/ModalReservationExtraPaymentRefund';
+import Organization from '../../../../../props/models/Organization';
+import Reservation from '../../../../../props/models/Reservation';
+import ExtraPayment from '../../../../../props/models/ExtraPayment';
+import KeyValueItem from '../../../../elements/key-value/KeyValueItem';
+import useSetProgress from '../../../../../hooks/useSetProgress';
+import usePushSuccess from '../../../../../hooks/usePushSuccess';
+import useProductReservationService from '../../../../../services/ProductReservationService';
+import useOpenModal from '../../../../../hooks/useOpenModal';
+import useEnvData from '../../../../../hooks/useEnvData';
+import useTranslate from '../../../../../hooks/useTranslate';
+import usePushApiError from '../../../../../hooks/usePushApiError';
+import Label from '../../../../elements/image_cropper/Label';
+import ModalReservationExtraPaymentRefund from '../../../../modals/ModalReservationExtraPaymentRefund';
+import FormPane from '../../../../elements/forms/elements/FormPane';
 
-export default function ReservationExtraPaymentDetailsCard({
+export default function ReservationExtraPaymentDetailsPane({
     payment,
     onUpdate,
     reservation,
@@ -60,36 +61,9 @@ export default function ReservationExtraPaymentDetailsCard({
     }, [onUpdate, openModal, reservation, organization]);
 
     return (
-        <div className="card">
-            <div className="card-header">
-                <div className="flex flex-grow card-title">Transactie details van de bijbetaling</div>
-
-                {isProvider && (
-                    <div className="card-header-filters">
-                        <div className="block block-inline-filters">
-                            {!reservation.canceled && (
-                                <button className="button button-primary button-sm" onClick={() => fetchExtraPayment()}>
-                                    <em className="mdi mdi-autorenew icon-start"></em>
-                                    Gegevens ophalen
-                                </button>
-                            )}
-
-                            {!payment.is_fully_refunded && payment.is_paid && (
-                                <button
-                                    className="button button-danger button-sm"
-                                    disabled={!payment.is_refundable}
-                                    onClick={() => refundExtraPayment()}>
-                                    <em className="mdi mdi-undo-variant icon-start"></em>
-                                    Bijbetaling terugbetalen
-                                </button>
-                            )}
-                        </div>
-                    </div>
-                )}
-            </div>
-
-            <div className="card-section">
-                <div className="card-block card-block-keyvalue">
+        <FormPane title={'Transactie details van de bijbetaling'} large={true}>
+            <div className="flex">
+                <div className="card-block card-block-keyvalue card-block-keyvalue-md card-block-keyvalue-text-sm">
                     <KeyValueItem label={translate('reservation.labels.status')}>
                         {!payment.is_fully_refunded && payment.is_paid && (
                             <Label type="success">{payment.state_locale}</Label>
@@ -125,6 +99,27 @@ export default function ReservationExtraPaymentDetailsCard({
                     <KeyValueItem label={translate('reservation.labels.method')}>{payment.method}</KeyValueItem>
                 </div>
             </div>
-        </div>
+
+            {isProvider && (
+                <div className="button-group">
+                    {!reservation.canceled && (
+                        <button className="button button-primary button-sm" onClick={() => fetchExtraPayment()}>
+                            <em className="mdi mdi-autorenew icon-start"></em>
+                            Gegevens ophalen
+                        </button>
+                    )}
+
+                    {!payment.is_fully_refunded && payment.is_paid && (
+                        <button
+                            className="button button-danger button-sm"
+                            disabled={!payment.is_refundable}
+                            onClick={() => refundExtraPayment()}>
+                            <em className="mdi mdi-undo-variant icon-start"></em>
+                            Bijbetaling terugbetalen
+                        </button>
+                    )}
+                </div>
+            )}
+        </FormPane>
     );
 }
