+++ react/src/dashboard/components/pages/transactions-view/TransactionsView.tsx
@@ -7,13 +7,15 @@ import useEnvData from '../../../hooks/useEnvData';
 import LoadingCard from '../../elements/loading-card/LoadingCard';
 import StateNavLink from '../../../modules/state_router/StateNavLink';
 import { useParams } from 'react-router';
-import TransactionDetailCards from './elements/TransactionDetailCards';
-import ReservationExtraPaymentDetailsCard from '../reservations-view/elements/ReservationExtraPaymentDetailsCard';
+import TransactionDetailsPane from './elements/panes/TransactionDetailsPane';
+import ReservationExtraPaymentDetailsPane from '../reservations-view/elements/panes/ReservationExtraPaymentDetailsPane';
 import useTranslate from '../../../hooks/useTranslate';
+import useAssetUrl from '../../../hooks/useAssetUrl';
 
 export default function TransactionsView() {
     const { address } = useParams();
 
+    const assetUrl = useAssetUrl();
     const envData = useEnvData();
     const translate = useTranslate();
     const setProgress = useSetProgress();
@@ -65,20 +67,63 @@ export default function TransactionsView() {
                 <div className="breadcrumb-item active">{'#' + transaction.id}</div>
             </div>
 
-            <TransactionDetailCards
-                transaction={transaction}
-                setTransaction={setTransaction}
-                showReservationPageButton={true}
-                showAmount={true}
-            />
+            <div className="block block-transaction-details">
+                <div className="card card-wrapped">
+                    <div className="card-section">
+                        <div className="card-title">
+                            <strong>{transaction.amount_locale}</strong>
+                            &nbsp;&nbsp;
+                            <strong className="text-primary">Bedrag</strong>
+                            <div className="flex flex-grow" />
+                        </div>
+                    </div>
+                    {transaction.notes && transaction.notes.length != 0 && (
+                        <div className="card-section">
+                            <div className="card-block card-block-notes">
+                                {transaction.notes.map((note) => (
+                                    <div className="note-item" key={note.id}>
+                                        <img
+                                            alt={''}
+                                            className="note-item-icon"
+                                            src={assetUrl(`/assets/img/note-icons/note-icon-${note.icon}.jpg`)}
+                                        />
+                                        {note.message && <div className="note-item-text">{note.message}</div>}
+                                        <div className="note-item-sign flex">
+                                            <span>
+                                                {'By ' +
+                                                    (note.group == 'sponsor'
+                                                        ? transaction.fund.organization_name
+                                                        : transaction.organization.name)}
+                                            </span>
+                                            <em className="mdi mdi-circle" />
+                                            <span>{note.created_at_locale}</span>
+                                        </div>
+                                    </div>
+                                ))}
+                            </div>
+                        </div>
+                    )}
 
-            {transaction?.reservation?.extra_payment && (
-                <ReservationExtraPaymentDetailsCard
-                    organization={activeOrganization}
-                    reservation={transaction.reservation}
-                    payment={transaction.reservation.extra_payment}
-                />
-            )}
+                    <div className="card-section form">
+                        <div className="flex flex-gap flex-vertical form">
+                            <TransactionDetailsPane
+                                transaction={transaction}
+                                setTransaction={setTransaction}
+                                showReservationPageButton={true}
+                                showAmount={true}
+                            />
+
+                            {transaction?.reservation?.extra_payment && (
+                                <ReservationExtraPaymentDetailsPane
+                                    organization={activeOrganization}
+                                    reservation={transaction.reservation}
+                                    payment={transaction.reservation.extra_payment}
+                                />
+                            )}
+                        </div>
+                    </div>
+                </div>
+            </div>
         </Fragment>
     );
 }
